Refactor: 공통 함수 및 컴포넌트 분리
